[chore] #41 패키징 추가 진행 group -> everyGroup, onceGroup으로 가도록!
공통으로 사용되는 vo, 클래스가 나와 패키지 변경 진행
