+++ src/main/java/com/ggang/be/api/comment/dto/ReadCommentResponse.java
@@ -1,7 +1,7 @@
 package com.ggang.be.api.comment.dto;
 
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
-import com.ggang.be.domain.vo.ReadCommentGroup;
+import com.ggang.be.domain.group.vo.ReadCommentGroup;
 
 public record ReadCommentResponse(@JsonUnwrapped ReadCommentGroup readCommentGroup) {
