+++ src/main/java/com/ggang/be/api/group/everyGroup/service/EveryGroupService.java
@@ -1,10 +1,10 @@
 package com.ggang.be.api.group.everyGroup.service;
 
 import com.ggang.be.domain.comment.CommentEntity;
-import com.ggang.be.domain.everyGroup.EveryGroupEntity;
-import com.ggang.be.domain.everyGroup.dto.EveryGroupDto;
+import com.ggang.be.domain.group.everyGroup.EveryGroupEntity;
+import com.ggang.be.domain.group.everyGroup.dto.EveryGroupDto;
 import com.ggang.be.domain.user.UserEntity;
-import com.ggang.be.domain.vo.ReadCommentGroup;
+import com.ggang.be.domain.group.vo.ReadCommentGroup;
 
 public interface EveryGroupService {
     EveryGroupDto getEveryGroupDetail(final long groupId, UserEntity userEntity);
