+++ src/main/java/com/ggang/be/api/mapper/GroupResponseMapper.java
@@ -2,8 +2,8 @@ package com.ggang.be.api.mapper;
 
 import com.ggang.be.api.facade.GroupType;
 import com.ggang.be.api.group.dto.GroupResponseDto;
-import com.ggang.be.domain.everyGroup.dto.EveryGroupDto;
-import com.ggang.be.domain.onceGroup.dto.OnceGroupDto;
+import com.ggang.be.domain.group.everyGroup.dto.EveryGroupDto;
+import com.ggang.be.domain.group.onceGroup.dto.OnceGroupDto;
 
 public record GroupResponseMapper() {
