+++ src/main/java/com/ggang/be/domain/group/everyGroup/application/EveryGroupServiceImpl.java
@@ -1,16 +1,16 @@
-package com.ggang.be.domain.everyGroup.application;
+package com.ggang.be.domain.group.everyGroup.application;
 
 import com.ggang.be.api.common.ResponseError;
 import com.ggang.be.api.exception.GongBaekException;
 import com.ggang.be.api.group.everyGroup.service.EveryGroupService;
 import com.ggang.be.domain.comment.CommentEntity;
-import com.ggang.be.domain.everyGroup.EveryGroupEntity;
-import com.ggang.be.domain.everyGroup.dto.EveryGroupDto;
-import com.ggang.be.domain.everyGroup.infra.EveryGroupRepository;
+import com.ggang.be.domain.group.GroupCommentVoMaker;
+import com.ggang.be.domain.group.everyGroup.EveryGroupEntity;
+import com.ggang.be.domain.group.everyGroup.dto.EveryGroupDto;
+import com.ggang.be.domain.group.everyGroup.infra.EveryGroupRepository;
+import com.ggang.be.domain.group.vo.GroupCommentVo;
+import com.ggang.be.domain.group.vo.ReadCommentGroup;
 import com.ggang.be.domain.user.UserEntity;
-import com.ggang.be.domain.vo.GroupCommentVo;
-import com.ggang.be.domain.vo.ReadCommentGroup;
-import java.util.Comparator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +22,9 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 @Slf4j
 public class EveryGroupServiceImpl implements EveryGroupService {
+
     private final EveryGroupRepository everyGroupRepository;
+    private final GroupCommentVoMaker groupCommentVoMaker;
 
     @Override
     public EveryGroupDto getEveryGroupDetail(final long groupId, final UserEntity userEntity) {
@@ -31,7 +33,7 @@ public class EveryGroupServiceImpl implements EveryGroupService {
     }
 
     @Override
-    public long getEveryGroupRegisterUserId(final long groupId){
+    public long getEveryGroupRegisterUserId(final long groupId) {
         EveryGroupEntity entity = findIdOrThrow(groupId);
         return entity.getUserEntity().getId();
     }
@@ -39,14 +41,15 @@ public class EveryGroupServiceImpl implements EveryGroupService {
     @Override
     public EveryGroupEntity findEveryGroupEntityByGroupId(long groupId) {
         return everyGroupRepository.findById(groupId).orElseThrow(
-                () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
+            () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
         );
     }
 
     @Override
     @Transactional
     public void writeCommentInGroup(CommentEntity commentEntity, final long groupId) {
-        everyGroupRepository.findById(groupId).orElseThrow(() -> new GongBaekException(ResponseError.GROUP_NOT_FOUND))
+        everyGroupRepository.findById(groupId)
+            .orElseThrow(() -> new GongBaekException(ResponseError.GROUP_NOT_FOUND))
             .addComment(commentEntity);
         log.info("everyGroupEntity add Comment success CommentId was  : {}", commentEntity.getId());
     }
@@ -62,18 +65,17 @@ public class EveryGroupServiceImpl implements EveryGroupService {
 
         int commentCount = commentEntities.size();
 
-        List<GroupCommentVo> onceGroupCommentVos = commentEntities.stream()
-            .sorted(Comparator.comparing(CommentEntity::getCreatedAt).reversed())
-            .map(c -> GroupCommentVo.ofEveryGroup(everyGroupEntity, c))
-            .toList();
+        List<GroupCommentVo> onceGroupCommentVos = groupCommentVoMaker.makeByEveryGroup(
+            commentEntities, everyGroupEntity);
 
         return ReadCommentGroup.of(commentCount, onceGroupCommentVos);
 
     }
 
-    private EveryGroupEntity findIdOrThrow(final long groupId){
+
+    private EveryGroupEntity findIdOrThrow(final long groupId) {
         return everyGroupRepository.findById(groupId).orElseThrow(
-                () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
+            () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
         );
     }
