+++ src/main/java/com/ggang/be/domain/group/everyGroup/dto/EveryGroupDto.java
@@ -1,12 +1,11 @@
-package com.ggang.be.domain.onceGroup.dto;
+package com.ggang.be.domain.group.everyGroup.dto;
 
-import com.ggang.be.domain.constant.WeekDate;
-import com.ggang.be.domain.onceGroup.OnceGroupEntity;
+import com.ggang.be.domain.group.everyGroup.EveryGroupEntity;
 import com.ggang.be.domain.user.UserEntity;
 
-public record OnceGroupDto(
+public record EveryGroupDto(
         long groupId,
-        String groupTitle,
+        String title,
         String location,
         String status,
         int currentPeopleCount,
@@ -17,12 +16,11 @@ public record OnceGroupDto(
         String category,
         int coverImg,
         String weekDay,
-        String weekDate,
         double startTime,
         double endTime
 ) {
-    public static OnceGroupDto toDto(OnceGroupEntity entity, UserEntity currentUser) {
-        return new OnceGroupDto(
+    public static EveryGroupDto toDto(EveryGroupEntity entity, UserEntity currentUser) {
+        return new EveryGroupDto(
                 entity.getId(),
                 entity.getTitle(),
                 entity.getLocation(),
@@ -34,11 +32,10 @@ public record OnceGroupDto(
                 entity.getIntroduction(),
                 entity.getCategory().toString(),
                 entity.getCoverImg(),
-                WeekDate.fromDayOfWeek(entity.getGroupDate().getDayOfWeek()).name(),
-                entity.getGroupDate().toString(),
+                entity.getWeekDate().toString(),
                 entity.getStartTime(),
                 entity.getEndTime()
-        );
+                );
     }
 
 }
