+++ src/main/java/com/ggang/be/domain/group/everyGroup/infra/EveryGroupRepository.java
@@ -1,6 +1,6 @@
-package com.ggang.be.domain.everyGroup.infra;
+package com.ggang.be.domain.group.everyGroup.infra;
 
-import com.ggang.be.domain.everyGroup.EveryGroupEntity;
+import com.ggang.be.domain.group.everyGroup.EveryGroupEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EveryGroupRepository extends JpaRepository<EveryGroupEntity, Long> {
