+++ src/main/java/com/ggang/be/domain/group/onceGroup/application/OnceGroupServiceImpl.java
@@ -1,16 +1,16 @@
-package com.ggang.be.domain.onceGroup.application;
+package com.ggang.be.domain.group.onceGroup.application;
 
 import com.ggang.be.api.common.ResponseError;
 import com.ggang.be.api.exception.GongBaekException;
 import com.ggang.be.api.group.onceGroup.service.OnceGroupService;
 import com.ggang.be.domain.comment.CommentEntity;
-import com.ggang.be.domain.onceGroup.OnceGroupEntity;
-import com.ggang.be.domain.onceGroup.dto.OnceGroupDto;
-import com.ggang.be.domain.onceGroup.infra.OnceGroupRepository;
+import com.ggang.be.domain.group.GroupCommentVoMaker;
+import com.ggang.be.domain.group.onceGroup.OnceGroupEntity;
+import com.ggang.be.domain.group.onceGroup.dto.OnceGroupDto;
+import com.ggang.be.domain.group.onceGroup.infra.OnceGroupRepository;
+import com.ggang.be.domain.group.vo.GroupCommentVo;
+import com.ggang.be.domain.group.vo.ReadCommentGroup;
 import com.ggang.be.domain.user.UserEntity;
-import com.ggang.be.domain.vo.GroupCommentVo;
-import com.ggang.be.domain.vo.ReadCommentGroup;
-import java.util.Comparator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,6 +24,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class OnceGroupServiceImpl implements OnceGroupService {
 
     private final OnceGroupRepository onceGroupRepository;
+    private final GroupCommentVoMaker groupCommentVoMaker;
 
     @Override
     public OnceGroupDto getOnceGroupDetail(final long groupId, final UserEntity userEntity) {
@@ -32,7 +33,7 @@ public class OnceGroupServiceImpl implements OnceGroupService {
     }
 
     @Override
-    public long getOnceGroupRegisterUserId(final long groupId){
+    public long getOnceGroupRegisterUserId(final long groupId) {
         OnceGroupEntity entity = findIdOrThrow(groupId);
         return entity.getUserEntity().getId();
     }
@@ -40,14 +41,15 @@ public class OnceGroupServiceImpl implements OnceGroupService {
     @Override
     public OnceGroupEntity findOnceGroupEntityByGroupId(long groupId) {
         return onceGroupRepository.findById(groupId).orElseThrow(
-                () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
+            () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
         );
     }
 
     @Override
     @Transactional
     public void writeCommentInGroup(CommentEntity commentEntity, final long groupId) {
-        onceGroupRepository.findById(groupId).orElseThrow(() -> new GongBaekException(ResponseError.GROUP_NOT_FOUND))
+        onceGroupRepository.findById(groupId)
+            .orElseThrow(() -> new GongBaekException(ResponseError.GROUP_NOT_FOUND))
             .addComment(commentEntity);
         log.info("onceGroupEntity add Comment success CommentId was  : {}", commentEntity.getId());
     }
@@ -63,17 +65,15 @@ public class OnceGroupServiceImpl implements OnceGroupService {
 
         int commentCount = commentEntities.size();
 
-        List<GroupCommentVo> onceGroupCommentVos = commentEntities.stream()
-            .sorted(Comparator.comparing(CommentEntity::getCreatedAt).reversed())
-            .map(c -> GroupCommentVo.ofOnceGroup(onceGroupEntity, c))
-            .toList();
+        List<GroupCommentVo> onceGroupCommentVos = groupCommentVoMaker.makeByOnceGroup(
+            commentEntities, onceGroupEntity);
 
         return ReadCommentGroup.of(commentCount, onceGroupCommentVos);
     }
 
-    private OnceGroupEntity findIdOrThrow(final long groupId){
+    private OnceGroupEntity findIdOrThrow(final long groupId) {
         return onceGroupRepository.findById(groupId).orElseThrow(
-                () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
+            () -> new GongBaekException(ResponseError.GROUP_NOT_FOUND)
         );
     }
 }
