+++ src/main/java/com/ggang/be/domain/group/onceGroup/infra/OnceGroupRepository.java
@@ -1,6 +1,6 @@
-package com.ggang.be.domain.onceGroup.infra;
+package com.ggang.be.domain.group.onceGroup.infra;
 
-import com.ggang.be.domain.onceGroup.OnceGroupEntity;
+import com.ggang.be.domain.group.onceGroup.OnceGroupEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OnceGroupRepository extends JpaRepository<OnceGroupEntity, Long> {
